+++ CHANGELOG.md
@@ -46,6 +46,8 @@ All notable changes to this project will be documented in this file, in reverse 
 
 - [#32](https://github.com/laminas/laminas-validator/pull/32) fixes PHP 7.4 compatibility.
 
+- [#34](https://github.com/laminas/laminas-validator/pull/34) fixes hostname validation for domain parts with 2+ dashes and with dash at the end.
+
 ## 2.13.0 - 2019-12-27
 
 ### Added